+++ src/main/scala/mmarquee/demo/AutomationTest.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package mmarquee.automation.Scala
+package mmarquee.demo
 
 object AutomationTest {
