+++ src/main/scala/mmarquee/demo/PacketLoggerTest.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package mmarquee.automation.Scala
+package mmarquee.demo
 
 object PacketLoggerTest {
